+++ src/GeneratorV2/GeneratorV2.csproj
@@ -9,4 +9,8 @@
     <StartupObject>GeneratorV2.Program</StartupObject>
   </PropertyGroup>
 
+  <ItemGroup>
+    <Folder Include="Process\Overloaders" />
+  </ItemGroup>
+
 </Project>
